+++ .github/workflows/ci-cd.yml
@@ -26,7 +26,7 @@ jobs:
       run: poetry install
     # Step 6. Run tests for package
     - name: Test with pytest
-      run: poetry run pytest tests/ --cov=straPy --cov-report=xml
+      run: poetry run pytest tests/ --cov=strapPy --cov-report=xml
     # Step 7. Use Codecov to track coverage
     - uses: codecov/codecov-action@v2
       with:
@@ -71,7 +71,7 @@ jobs:
           pip install \
           --index-url https://test.pypi.org/simple/ \
           --extra-index-url https://pypi.org/simple \
-          straPy
+          strapPy
     # Step 7. Publish to PyPI
     - uses: pypa/gh-action-pypi-publish@release/v1
       with:
