+++ docs/Makefile
@@ -3,7 +3,7 @@
 # You can set these variables from the command line.
 SPHINXOPTS    =
 SPHINXBUILD   = python -msphinx
-SPHINXPROJ    = straPy
+SPHINXPROJ    = strapPy
 SOURCEDIR     = .
 BUILDDIR      = _build
