+++ docs/conf.py
@@ -6,7 +6,7 @@
 
 # -- Project information -----------------------------------------------------
 
-project = u"straPy"
+project = u"strapPy"
 copyright = u"2022, Julien Gordon, Gautham Pughazhendhi, Zack Tang, Margot Vore"
 author = u"Julien Gordon, Gautham Pughazhendhi, Zack Tang, Margot Vore"
